JrtFinder: detect if the JrtFileSystemProvider exists, and if not, use the lib/jrt-fs.jar to load anyways
The `.addRuntime(StructContext)` method now also handles the case without a jrt:/ provider, by looking in directory specified by the system property `java.home`.
The `.addRuntime(StructContext, File)` method checks if the runtime to add has the `lib/jrt-fs.jar` file, and if that file exists and the currently executing java has a jrt:/ provider, then that provider is used to load the java from the given `javaHome`.
If the currently executing java has no such jrt:/ provider, the file `lib/jrt-fs.jar` is opened and loaded via an `URLClassLoader` to create a new `FileSystem` for loading the java located at `javaHome`.
If neither of these case happens, it is assumed to be a runtime in the old (Java 8 and lower) format, so the `lib` and `jre/lib` folders are searched for `.jar` files.

# Potential Security issues:
If such a file `lib/jrt-fs.jar` is from a malicious attacker, they can get VineFlower to execute the attackers code.