Remove more unnecessary java.lang.module usage, where just a normal String would be enough
Since the ModuleDescriptor passed to the ModuleBasedContextSource was only used to get the nameAndVersion from it, replacing that with a String allows removal of more of the java.lang.module api.
For the ModuleContextSource, it's also better to pass the ModuleReader and the nameAndVersion string in directly. This also gives the option to also log the version of the module in case something goes wrong.